[od6.0] Two fault display-only-at-sections update issues (part of OD-285)
Summary:
1) Using v-shortcut to display-only-at-sections, and next in tree uncheck
display only-at-sections does not redisplay fault planes. Reason is that the
v-shortcut also displays-only-at-horizons (revision r41072/D8834 by Nanne),
which needs to be undone in the tree as well in that case. It would require
all kind of extra logic to improve this, but discussion with Ayrat & Farrukh
learned that it is actually undesirable that the v-shortcut also activates
display-only-at-horizons. Main reason is that the algorithms behind it still
show a bad performance, so the display-only-at-horizons should be kept fully
independent.
(2) Fault sticks, editing nodes and active stick were not hidden in case of
display-only-at-horizons (without display-only-at-sections at same time).
These elements are obstructing and also unnecessary because picking on horizons is not supported.

Test Plan:
1) Using v-shortcut to display-only-at-sections, and next in tree uncheck
display only-at-sections does now redisplay fault planes again, and horizon
intersections are merely shown when explicitly checked in tree. 
(2) Fault sticks, editing nodes and active stick are no longer obstructing
when merely displaying-only-at-horizons, but do return on sections if also
selecting display-only-at-sections. Tested with Fault-A and Demo_1->MFS4 in
F3_Demo. Own picked two-stick fault refuses to show intersection line on
horizon. Separate investigation required.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D10132