od6.0: Fixed problem with right-click popup menu
Summary:
There's a problem with the bounding box of text items. It becomes really large when zooming in. A right-click in the basemap is often a right-click on the text in stead of on the 2D line (depending on the order added to the basemap). This change does not fix the bounding box but works for now in the 2D case. The text is now 1 z-level lower than the lines themselves.
Fixes OD-297

Test Plan: Added many 2D lines to the basemap. A right-click now always pops-up for the correct line.

Reviewers: ranojay

Reviewed By: ranojay

Differential Revision: https://phabricator.dgbes.com:8443/D10197