od6.0: OD-409, [CR 20160406175222] having problem uploading well tops
Summary: While changing the name using the virtual keyboard the cell was deleted once the value of the cell was set to null. This has been fixed. Now the cell content will be cleared and cell wont be deleted

Test Plan: It has been tested using well F02-1 and the renaming is working fine.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D10157