od6.0: added possibility to change X values of MathFunction
Summary:
in HCProcessing one single function is used, with a fixed number of samples and Y values always equal to the Z of the trace sample. What changes from trace to trace is the X-vector, always monotonic, but with other values. Thus this function is needed to avoid data allocations from trace to trace.
Requested for OD-483

Test Plan: wrote a new 3DHorizonCube. Got expected results.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D10211