od6.0 : Current Interpreted object is now remembered in OD to be utilised in various places
Summary: This information is the passed on from 3D scene to 2D viewer to highlight the last interpreted object. OD-401. Also made small fixes for individual treeitems so that interpretation can be seamless

Test Plan: Tested with multiple types of interpretable objects at the same time. The last selected tree object in the tree is remembered & passed on to the 2D viewer

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D10229