[od] Window ordering swapped for multiple scenes with fault(s) (OD-366)
Summary:
Window ordering swapped for multiple scenes with fault(s). Happened if fault
item was selected or deselected by clicking either in data tree or in scene,
while the focus window in workspace was not the top left one at that moment.
There is a relationship with the fault toolbar being shown or hidden during
this action. It looks like QMdiArea does not obey its WindowOrder in that case
and uses ActivationHistoryOrder instead of default CreationOrder over there.
The solution is based on the observation that (1) the effect disappears after
Scene->Tile->Horizontal/Vertical and returns after Scene->Tile->Auto, and (2)
upto OD4.6 the newly added second scene was displayed right instead of left.
Apparently, the tileHorizontal/Vertical() functions do not only perform a
resize but (by accident) also a reorder operation. Adopting this reorder
operation in the tile() function as well seems to do the job.

Test Plan:
Tested with up to five scenes in F3_Demo. Scenes stay nicely sorted top-down
and left-right now. Tried Fault_A (in turns) in any of the scenes and next
clicked in all scenes and trees in order to force a window swap. It did not
happen anymore. Also verified that Scene->Cascade does not have a similar
problem.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D10253