[OD]OD-567 Manual Draw: manually picking a segment and moving to another inline without double-click results in unexpected behaviour
Summary:
The mentioned issue in the jira report is 
if we done a line drawn by manually single click without doing double click, and then moving plane data display by press keyboard. 
In this case, the user expect the un-ended line should be automatically end ( added to horizon ).

This solution added a planeMovedEvent notifer and implement it in uiMPEman.

Test Plan:
Tested in F3-demo by

1.doing pre-load,
2.load inline
3.open new horizon by auto and maual
4.set manul mode in tracking setup
5.pick several picks don't doing double click and then moving inline by press "F", the green line becomes red ( ended )
6. repeat many times. works fine.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D10219