[OD]OD-193 Implement Ctrl+Left-click+drag eraser in Section Auto-track mode
Summary:
The key point in this solution, we need catch the control button state information for every picks during sowing action. Then we can handle with 
picks in different mode ( erase or new seed ). In default sowing situation, the visssow changes the button state to default mask( leftbutton ). 
So I do particularly reset the sequentmask as( left button and control button ) for horizon in visemobjectdisplay. and then the rests of code are made follow 
this change in horizon tracking.

Test Plan:
Tested in F3-demo by
1. load inline or randomline or 2d line
2. start new 3d horizon or 2d horizon.
3. set hoizon track mode as "auto" in setup dialog.
4. pick or sowing some picks.
5. do ctrl + sowing under existing horizon, you will see existing horizon is erased.
6. do test as usually in manual mode. it work as usual.

Tests working fine in both of 2d and 3d horizon

Reviewers: nanne, jaap

Reviewed By: jaap

Differential Revision: https://phabricator.dgbes.com:8443/D10223