[OD]Possible horizon crash bug fixing in BinID manipulation
Summary:
don't know if this revision related to some Jira issues or not.
This revision is created on the report from Intel Parallel Inspector which found two threads manipulating BinID
Thread 1: caused by actioncalled.trigger and actionfinished.trigger in function trackingFinishedCB in mpeengin.cc. it leads to call function
updateAuxData() in vishorizondisplay. in that function it will delete catched BinID
Thread 2: caused by emseedpicker which leads to call emobject's setPos function in which it will trigger the callback emobjectCB in vishorizondisplay, and then 
then function setRandomPosData will be called. in the setTexture function, getBinID() is called. So one thread is trying delete it, another thread is trying
to get value from it.
The revision added two lockers

Test Plan: Tested with loading horizon and auto tracking horizon, it works as usual without slowly the speed.

Reviewers: nanne

Reviewed By: nanne

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D10257