od6.0: OD-220, 564
Summary:
1. The checkshot corrections were not including the shallow and deep point
2. In apply checkshot window the green curve was not resizing upon maximixation

Test Plan: All the points are included in the model as desired, The check shot data is displayed in the use editor window, picked up point integrated with d2t model are displayed for automatic method and original d2t for the none method. Green curve is maximising now.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D10273