od6.0 : Fix for multiple updates of color table mapper ranges during bitmap drawing
Summary:
https://jira.dgbes.com:8443/browse/OD-307
Problem : When ColTab::MapperSetup::Type == Auto then everytime a bitmap is drawn the scaling function is called leading to slightly different ranges every time. Also the ranges which are user to draw the botmap is not updated in the uiColorTable. Instead uiBitMapDisplay::getDataRange() is called which gives another set of range which is set to the uiColorTable

Solution: When Auto is set from anywhere in the UI or datapack is changed the the ColTab::MapperSetup::range_ is set to undefined. Only when it is undefined, then scaling function is called. Once scaling function gives back range it is set back & then until & unless paramters are not changed range is not recalculated.

Test Plan: Tested in Basemap. Horizon display problem solved. Tested in 2D Viewers displaying attributes with saved colr settings & non saved color settings. Ranges not updated dynamically. Tested in PreStack 2D Viewers, SynthRock viewers, WCP, PDF display

Reviewers: kristofer, mahant, raman

Reviewed By: raman

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D10271