od6.0: Fix for MATLAB step multiinput
Summary:
Added a uiGroup to serve as parent for the uiTable objecs in the MATLAB step. This solved the problem of non-responding comboboxes and other cells in the tables.
See OD-447

Test Plan:
Setup a MATLAB environment on my linux pc with some runtime libraries. The table works OK now.
Needs tests in real release.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D10333