od6.0: Fixed a Desc update problem
Summary:
As in OD-559, if we create a volume with 3 components first, the display reload will crash if we create a same name volume with less components. It turn out the Desc is not updated.

Please check if this is the right place to update the Desc.

Test Plan: Followed OD-559, the crash is fixed

Reviewers: helene

Reviewed By: helene

Differential Revision: https://phabricator.dgbes.com:8443/D10360