od: Have trace interpolator use 1e30 as undef value
Summary:
The default interpolator in SeisTrc has 0 as its undef value. This gives problems when extracting stored data (with undefined values) on horizons. The resulting attribute display had zeros where undefs were expected.
See OD-435 for a detailed description of the workflow. The workflow is HorizonCube3D slider related, but the attribute extraction for horizoncube events is identical to the one of normal 3D horizons.
I could have changed the undef value in seistrc.cc, but I'm not sure if that will cause problems elsewhere where a zero undef value is assumed.

Test Plan: Applied the TFL depth volume to a 3D horizon. Undef values can now be seen. Also applied this same volume to the 3D slider with nice results.

Reviewers: helene

Reviewed By: helene

Differential Revision: https://phabricator.dgbes.com:8443/D10362