[od6.0] Horizon sorting slows down multiple horizon loading (part of OD-386). PORTED FROM [od] WITH datastamp_ HIDDEN PARAMETER HANDLING
Summary:
Horizon sorting slows down multiple horizon loading (part of OD-386)
(1) Sorting was done after adding each horizon to the tree, where doing it
after all selected horizons have been added is sufficient.
(2) Sorting itself was very slow in case the survey contains many horizons.
The bottleneck is file system access to check the date of last change for
all horizons. Added a static function sortHorizons(.), which only does this
check for horizons that are actually in the tree. Now sorting takes only a
fraction of the time spend for loading and constructing a horizon.
(3) The horizonrelations.txt file of survey F3_Demo_CSM in /dsk/d46/ODData
contains 1071 3D Horizon entries. A hard-coded limit in the read(.) function
was taking into account only the first 1024. Increased this to 1,000,000
for safety, although no limit at all seems also fine given the current break
condition in the loop.
(4) After fixing issue (3), horizon sorting ended up in a kind of live-lock
while handling entry 3D Horizon.1046. Investigation showed that the hasChild(.)
function is busy searching a "virtual" RelationTree of about 10^24 nodes,
because the same nodes show up in many different branches. Although this
might be the result of unexpected behaviour from the algorithm producing the
horizonrelations.txt file, there is just no sense in having the hasChild(.)
function search the same "descendant" argument in the same node over and over
again. Note that it is not the lack of recursion in the new implementation
that solves this case, but the isPresent(.) check.
(5) Added wait-cursors during horizon depth attribute texture generation
and also during horizon sorting in data tree.
(6) [unrelated] Old implementation of setSectionDisplayRestoreForAllHors(.)
function was inefficient and really slowed down horizon loading in case the
scene contained a huge amount of DataObjects, e.g. seed markers of faults.

Test Plan:
Testing with F3_Demo_CSM shows that horizon loading has become "scalable"
again. Execution time is near-linear with the number of horizons loaded.
Now almost all time is spent on horizon loading + construction, and
texture generation. All algorithmic changes are supposed to be invariant in
their input-output behaviour, and only make things faster. I tried to be
very precise in this, and did not discover any change in behaviour so far.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D10381