[OD]OD-538 Polygon picks undoing problem
Summary:
The problem comes from the first undo location coordinate which was defined as (0,0,0) when clicking on the previous pick( that means the pick can be found from existing surface).
So the initial cooedinate of undo location should be undefined not 0. Then we will follow this idea to handle with other undo operation.

The changes in this revision

Test Plan:
No way to test it at this moment, because Bert removed undo/redo codes temporal, he is doing the redesign of undo/redo mechanism in od.
But it is tested in od6.0 using same code. it works fine. So this change should be re-tested after Bert finish his work.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D10328