od6.0: SEG-Y scan for survey setup: limit number of traces to be scanned for quick scan. [OD-636]
Summary: When trying to use a SEG-Y file to set up a survey, wizard tried to scan until it found at least one inline and one crossline change. With bad bytes or missing data it would scan the entire file. Now, limiting this to 10000 traces. This takes about 10 seconds for the mentioned SEG-Y files over our network.

Test Plan: Did just a few tests, low risk as the while ( true ) is changed to a limited for loop.

Reviewers: helene

Reviewed By: helene

Differential Revision: https://phabricator.dgbes.com:8443/D10391