od6.0: OD-448 PreStack gradient attribute computation is producing results with sign reversal + some minor GUI issues
Summary:
This is reopened issue. The sign reversal which was due to degrees and radian conversion bug has already been fixed. Few GUI issues have been in this revision:
1. The range field is visible now. It will be grayed out when compute angle check box is checked.

2. When the Calculation type is AVO, Gather Type is Offset and compute angle is checked then the X range field will be set to null despite of the fact that it contains some values. This is because those values will be redundant in this case.

3.When Gather type is offset and calculation type is AVO then X Axis Transformation is set to None but not grayed out.

Test Plan: All the changes are working as desired. There has been no change in the logic. The attribute creation is not hampered.

Reviewers: raman

Reviewed By: raman

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D10431