od6.0: OD-700, Well-to-Seismic Tie: logs and synthetic trace below certain time/depth get cropped when using checkshot editor
Summary: The deeper part of the logs were truncated when checkshot corrections were applied and was identical to the non correction display. This happened only in the Use Editor mode. The bug was due to the fact the corrected values were read for automatic correction only and not for use editor correction.

Test Plan: The plots are now correctly generated. This has been confirmed by reading the values displayed at the status bar and found to be similar for both automatic and use editor case

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D10430