[od] Inline attribute in depth scene is reloaded when becoming deselected. WITH SIMPLIFICATIONS THAT ARE NOT IN [od6.0] FOR SAFETY.
Summary:
Inline attribute in depth scene is reloaded when becoming deselected,
e,g.  when clicking on crossline item. Caused by TrcKeyZSampling comparison
in isManipulated() function suffering from numerical instability in the
z-range domain. Current eps-makeshift only works for time scenes in time
surveys and depth scenes in depth surveys. Introduced new isEqual(tkzs,eps)
function, which takes its eps relative to the zstep if not defined explicitly.
For safety, the operator== will only request this relative eps in the trunk,
and keeps using the hard-coded absolute eps for time or depth surveys in the
stable.

Test Plan:
Inline attribute in depth scene no longer reloads when becoming deselected.
The only isManipulated() function with equivalent code is in the MPEDisplay
class. Did not test that one, because for the moment it is not used in OD.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D10482