od6.0: OD-716 OD-671, Well to seismic tie log display issues
Summary: The logs in the well to seismic tie were not correctly displayed and lacked a correlation between the reflectivity curve and the synthetic. This was because of the wrong bool values being read. Apart form this the z in feet button was not working properly. With this fix in 6.0, when unchecking the z in time button the laebl text for display changes to z in feet or z in meter depending upon the user choice in which unit he wants to see the logs. The z scale of the log has been modified and now diplay the value in feet when z in feet is selected. The values in the task bar also display correct unit now. Initial unit of display of logs is selected on the basis of value in the par file

Test Plan: The wells in the F3Demo have been checked and display is found to be corrrect, also the test survey gives positive result

Reviewers: raman

Reviewed By: raman

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D10484