od6.0: OD-654, Import SEG-Y does not work if executed in batch
Summary: The range selection i.e. whether to use Inl/Crl or XY was not read properly because wrong par file was being read. This lead to wrong calculation of the BinIDs and hence no data was imported in case of Batch Programming.

Test Plan: The import of segy file is now working properly. This has been tested in both F3Demo as well as in the test survey provided.

Reviewers: bert

Reviewed By: bert

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D10450