od6.0: Fix for vague error message [OD-529]
Summary: In OD5 if you restored a session with a 2D line and the data was no more available for that line, you got an error message from the reader. In od6 we get an error early on when we check for ranges in the Attribute Engine but the correct error message was not given. The new function checks for the avialability of 2D dataset if the output is desired for a 2D line.

Test Plan: Restored a session which contains a 2D line with a missing stored attribute. The new error message shows up instead of the old general error message.

Reviewers: helene

Reviewed By: helene

Differential Revision: https://phabricator.dgbes.com:8443/D10501