[OD] Bug fixing for set default z stretch to different scene part 1
Summary:
1. since we have changed the manipulation of z stretch in each scene, we need change the default z stretch too when create a new scene.
The z domain of a scene, its default definition is TWT. so in the setup function of vissurscene we can set its default z stretch, but for other scenes we don't 
know the ZDomain type during the stage of scen setup. Thus we have to set the z default value in the function setZAxisTranformation(). I checked this function, it is called only
once when create a new scene.

2. for wheeler scene, it is something different after do investigation. it will handled in another revision for plugin of UISSIS

Test Plan:
Tested with plugin revision together, it works fine in all scenes: we define the default z stretch for different type of scene, and then we can create different default scene
with different z stretch value.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D10485