od: Implemented a generic copier to transfer data between two 2D arrays
Summary:
as requested by Nanne for the Horizon Tracker. I used an implementation I already had in HC3DProcessor, but made it more generic and multi-threaded.
Instead using a TrcKeySamplingIterator I preferred to browse the lines, since it gives the opportunity to transfer all sampled of the line in one call providing that the 2D array has a pointer.

Test Plan: tested with size(in)>size(out) and two arrays of identical size. Made some tests in gdb, looks fine, however my application does not export the result to the database, thus not fully tested. Should be more easy to see from other applications.

Reviewers: nanne

Reviewed By: nanne

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D10526