od6.0: Forwarding the progress meter of the executor group to its executors
Summary:
the executors of the group do not inherit the progress meter of the group itself. This later is created while the execute() of the group is created, thus it is not possible to do the forwarding for instance by reimplementing the execute function.
To be on the safe side I could make this forwarding optional at least in od6.0, please advise.

Test Plan: tested in gdb that the children got their progressmeter from the class MultiSurfaceExtractor

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D10537