[od6.0] AuxData inherited when adding same 3D horizon in tree once more (OD-760) PORTED WITH HIDDEN PARAM CHANGES
Summary:
AuxData inherited when adding same 3D horizon in tree once more. Page-up/down
attributes even add as separate attributes. Both for coltab mapping and RGB
blending. This bug appeared after introduction of "tracking" AuxData. Fix only
adds AuxData of type Tracking to the new horizon tree and only if tracking is
possible directly. I did consider dirty fixes like "only for empty horizons",
but adding the new AuxDataType functionality is more flexible for the future.
Especially because it is unclear whether or not multi-type AuxData sets are
(going to be) allowed in the SurfaceAuxData class. And in which cases it is
actually sound to call removeAll(). One can think of many strange scenarios
having the same horizon object appear multiple times in the same or different
scene trees, while the user is allowed to add and remove horizon data at will
everywhere. However, let's focus on actually used workflows first.

Test Plan:
Add a horizon to the tree, add horizon data, add the same horizon again,
optionally with RGB blending, and the horizon data from the first is no
longer inherited by the second. Next add a new horizon to the tree. The
"tracking" AuxData attributes are still added automatically. Also if the new
horizon is saved in between tracking (or auto-saved in future), after which
it can also be added another time in the same or a different scene tree.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D10517