od6.0: OD-776, [CR 20160617045432] SeisCBVS2DLineGetter::nextStep()
Summary: The actual bug was not reproducible but the call stack could easily be reproduced. The call stack was produced when attribute window is opened. mIsUdf check was missing for nrbytes that were being passed.

Test Plan: Undefined values were passed and call stack was reproduced. A check for undefined values was placed and the crash did not occur.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D10531