[OD]od-590, 3D Auto-Tracking: display only at sections should apply to locked points texture (Show Locked)
Summary:
This revision is not only fixed the problem in od-590 but also some other issues not mentioned in Jira issue. for example,
the manipulation of combining "at section" and "lock",  "at section" and "lock show", part "locked show" and part "unlock", 
and display of plane data moving.

Test Plan: Tested with Aryat together in F3-demo working fine in all manipulations which meet the requirments from Aryat.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D10555