[OD6.0]OD-772 and OD-807 about edge of horizon title issues.
Summary:
This revision is ported from revision D10544
OD-807 is happened in doing horizon tracking when the inline or cross line are located at the last row or last column of one title. The algorithm in detect isolated line missed this situation.
OD-772 is happened in interpolating a horizon from low resolution to high resolution. In the algorithm of detectedisolatedline in the class vishortileresolutiondata,
it does detecting tracking line which is an independent line. that means its neighbour grid should have no data, otherwise it will be not a line but a part of triangle. The idea works in horizon tracking very well.
But when we do interpolating from (for example) the resolution 2 inline to 1 inline, the new line between previous tiles ( resolution as 2 ) is not detected. That is bug OD-772. The reason is the detectedisolatedline
algorithm think it is not independent line because its neighbours has data.

So for OD-772, I added an extra function to let the agorithm don't use neighbours when we do only interpolating horizo not do horizon tracking.

Test Plan: Tested following the description in Jira in F3-demo, they all working fine.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D10554