Added main() to testprog.h and let the test programs have testMain functions.
Summary:
If you call the ExitProgram from the main() function, valgrind will report
about memory leaks (as local variables are not deleted).

We here run the same plan as with the batch programs, that we have another
function being called (testMain()), and implement the main() in the
testprog.h headerfile.

Test Plan: Compiles fine. Some tests failed, but they failed without this change as well.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D10583