od6.0: Fix for undef gaps in 2D seismic display.
Summary: Made datapack to be created only for the TrcKeyZSampling of data to solve the issue. It is now possible to set data with a TrcKeyZSampling that is different from that of 2D line's geometry after Jaap's changes (Task: T254.) Jira Issue: https://jira.dgbes.com:8443/browse/OD-808.

Test Plan: Tested this in /dsk/d35/surveys/OD808 survey where data trace step is greater than that of geometry sampling step and also in F3_Demo.

Reviewers: jaap

Reviewed By: jaap

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D10580