[OD] Ensure several visualization callback function into main threads
Summary:
When doing auto tracking, or re-tracking, function HorizonTrackerMgr::taskFinished( CallBacker* ) or mpeActionFinishedCB and so on. These function are working on 
working threads which are going to call visualization functions to update the scene when the scene is been rendering in main threads ( we call visualization thread) .
This revision is trying to ensure to put emchangeCB and updateAuxData in horizon into main thread.

Test Plan:
Tested in F3-demo in auto tracking, re-tracking, remove seeds, pick seeds in 3d horizon tracking on randomline and inline. it works fine.
Tested also in 2dview by tracking 2d horizon. it works as normal.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D10489