[od6.0] Binary compatibility fix for reimplemented virtual functions in Ding's D10609.
Summary: Binary compatibility fix for reimplemented virtual functions in Ding's D10609.

Test Plan:
It compiles and the call stack in the debugger proves that the right functions
are called when using polygon selection tool in a 3D scene with a 2D horizon.
Also checked for all other functions added in D10609 whether they happen to
reimplement a virtual function.

Reviewers: nanne

Reviewed By: nanne

Subscribers: ding

Differential Revision: https://phabricator.dgbes.com:8443/D10626