od: Fixed some problems with pre-loading 2D data and the histogram
Summary:
In parallelreader, scaler should be cloned. In uiSeisPreloadSel, selectedGeomIDs was missing implementation. The creation of histogram needed a special impl for 2D. At the moment the histogram is based on the first selected 2D line.
Fixed bug in moving clipping lines in histogram.

Test Plan: Tested loading 2D data in F3 project and in nLog_2D. All seems to work fine again.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D10627