od6.0: OD-816, Unit selection issue when creating log from saved rock physics formula
Summary: The Rock Physics had option for converting the log units, but this value was set to default log unit value which made little sense since no conversion was required in this case. Final log units were set to SI units. Now the input log units are set to by default in SI unit. Warning is popped up if both input and final logs are in non SI unit and an error message is popped up if one of them is in non SI unit.

Test Plan: The logs are displaying correct unit. Warning and error messages are popped up as satisfied by criterion.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D10586