od : Support recall of CallBacker functions with capsule
Summary:
The CallBacker functions with Capsule has to be called again from QEventLoopReceiver. So I create a clone of the Capsule ( because the capsule gets deleted in the mean time ) & pass it on in the same manner as mEnsureExecutedInMainThread & then in the QEventLoopReceiver::event() I delete the Capsule. If any capsule callbacker is passed on mistakenly with mEnsureExecutedInMainThread call , then the program would crash anyway. This is a concept. If the methodology is acceptable then , we can try it out in other places. https://jira.dgbes.com:8443/browse/OD-862

Fix for lint error

Test Plan: Before yesterdays changes uiODVw2DPickSetParentTreeItem::picksetAddedCB was been called from non GUI thread. But now it is not called anyway. But according to yesterdays test things were working properly.

Reviewers: bert

Reviewed By: bert

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D10642