od6.0 : Fix for restoring saved AttributesDescSet in SynthRock CrossPlot
Summary: When a saved descset is opened in the SynthRock crossplot then the inputs were not set properly. So right now I am setting the first datapack to all the attributes. But a comprehensive solution would be using of uiStoredAttributeReplacer. But that can only be done after consultation with Helene. https://jira.dgbes.com:8443/browse/OD-865

Test Plan: The error messages are gone now. All the attributes have the first datapack as an input. But we cannot really implemement what Claire wants

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D10643