[OD6.0]OD-790 Horizon Auto-tracking: Locked color doesn't update immediately
Summary:
Actually the bug was fixed in revision 44773. Because revision 44789 was ported from a backup of file which was locally there for long time due to delayed review. and then the mistake was made by porting this file to revision 44789.
This revision has corrected the mistake to revision 44773.

Test Plan: Tested with F3-demo working fine

Reviewers: jaap

Reviewed By: jaap

Differential Revision: https://phabricator.dgbes.com:8443/D10664