[OD6.0]OD-877 While interpreting a horizon, at some point ctr+click starts removing neighboring seeds
Summary:
This revision is ported from revision D10690

This bug is only happened in horizon tracking using random line. The problem was caused by un-corrected index(Pos::Id) which corresponding the deleted node ( as a marker displayed
in the scene). The previous computation of find closest pos::id in the function getPosAttribPosID() is to find the closest marker. but some times the pos:id is not correct,
because the markers are not always keep the consistency with node lists, due to undefined node can be generated from horizon tracking using Randomline.

This solution uses node list instead of markerset in the function getPosAttribPosID().

Test Plan: Tested many times in F3-demo using random line doing horizon tracking. it works fine, without problem any more.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D10691