od6.0: OD-892 Save Z values in depth from a depth scene
Summary: Save as horizon data option was not provided for z values in any displayed. Now we can save the z values as horizon data in depth scene for time survey and as TWT for depth survey.

Test Plan: The changes have been checked using F3Demo2015 and Seg Salt 2 and the values are being correctly saved and later displayed in the original scene

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D10684