od6.0: OD-870, Well log units: '-' aka None unit is missing
Summary: In the well manager if a well log was selected with no dimension then instead of '-', seconds was displayed as unit. This has now been corrected. The setup for unit selection was not informed that there could be a dimensionless quantity and hence it was selecting seconds as default in this case

Test Plan: The unit of porosity was earlie displayed as second is now displayed as '-'. tested on well F02-1 in F3Demo-2015

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D10640