[OD]OD-807 When tracking a horizon with a step of 64 lines, it shows the interpretation only at sections
Summary:
Improved the algorithm detecting isolated line in horizon where the horizon is located at the edge of a tile.
since at the edge of tile, it is overlay part between two tiles, the previous method looking the neighbour tile has one row or col less.
This revision corrected this error and make it work properly.

Test Plan: Tested in F2-demo with Ayrat together. it is fine.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D10742