Rework of threaded NetreqConnection with event-loop.
Summary: Attempt to fix OD-691. Instead of letting waiting read-request poke the socket-thread, it is merely waiting for data to arrive. This makes a simpler threading function, but a sender class had to be implemented to send the data from within the event loop of the socket thread.

Test Plan: All test programs work as before.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D10759