od: Fix for cumulative sum wrong in two classes
Summary: when run on a float dataset, the result is incorrect because of roundoff errors. Fixes OD-961 and OD-999. Added up to four template variables to cover all cases and avoid the need to specialize thus duplicate most of the implementation of the parallel classes.

Test Plan: made a test program. Mode tests fail without the patch. Tested the unconformity tracker: Results are no longer CPU-dependent. Tested also briefly the frequency and spectral decomposition attributes.

Reviewers: kristofer

Reviewed By: kristofer

Subscribers: nanne, helene

Differential Revision: https://phabricator.dgbes.com:8443/D10760