od6.0: Fix for Hilbert transform when input is entirely undefined
Summary: fixed a division by zero when all input samples are undefined. Made it clear that the implementation relies on the presence of a pointer in both input and output ValueSeries. Added a protection for the developer to detect the case where all input samples are undefined (the function should not be called); For performance reasons this is only done in debug mode. Fixes OD-1049.

Test Plan: tested by computing a Dip-Steering volume with BG algorithm on Total dataset, where the process got stuck in an almost endless loop.

Reviewers: helene

Reviewed By: helene

Subscribers: nanne, kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D10793